:truck: Rename data_loaders to data
