+++ docs/src/api/data.md
@@ -1,0 +1,3 @@
+# Data Loaders
+
+::: icenet_mp.data
