--- docs/src/api/data_loaders.md
@@ -1,3 +1,0 @@
-# Data Loaders
-
-::: icenet_mp.data_loaders
