+++ docs/src/api/index.md
@@ -7,7 +7,7 @@ Select a module from the navigation to view its full API.
 |---|---|
 | [callbacks](callbacks.md) | Training callbacks (activation saving, etc.) |
 | [compatibility](compatibility.md) | Compatibility functions for external libraries |
-| [data_loaders](data_loaders.md) | PyTorch data loaders and datasets |
+| [data](data.md) | PyTorch data loaders and datasets |
 | [geotools](geotools.md) | Geospatial helpers (projections, grids) |
 | [ingestion](ingestion.md) | Data ingestion utilities |
 | [losses](losses.md) | Custom loss functions |
