+++ tests/data/test_combined_dataset.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from icenet_mp.data_loaders.combined_dataset import CombinedDataset
-from icenet_mp.data_loaders.single_dataset import SingleDataset
+from icenet_mp.data.combined_dataset import CombinedDataset
+from icenet_mp.data.single_dataset import SingleDataset
 
 
 class TestCombinedDataset:
