+++ tests/data/test_single_dataset.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from icenet_mp.data_loaders.single_dataset import SingleDataset
+from icenet_mp.data.single_dataset import SingleDataset
 from icenet_mp.types import DataSpace
